changed error handeling to return null and moved AssetObject classes into a sub directory
