+++ src/main/java/ResourceDiscovery/AssetObjects/InstanceCloudSqlObject.java
@@ -1,4 +1,4 @@
-package ResourceDiscovery;
+package ResourceDiscovery.AssetObjects;
 
 import java.util.Map;
