+++ src/main/java/ResourceDiscovery/AssetObjects/InstanceComputeObject.java
@@ -1,16 +1,17 @@
-package ResourceDiscovery;
+package ResourceDiscovery.AssetObjects;
 
 import java.util.Map;
 
-public class DiskComputeObject extends AssetObject {
-    public DiskComputeObject(Map<String,String> assetObjectsMap) {
+public class InstanceComputeObject extends AssetObject {
+    public InstanceComputeObject(Map<String,String> assetObjectsMap) {
         setFieldValues(assetObjectsMap);
     }
+
     public void setFieldValues(Map<String,String> itemsMap) {
         this.kind = itemsMap.get("kind");
         this.name = itemsMap.get("name");
         this.id = itemsMap.get("id");
-        this.type = getLastSeg(itemsMap.get("type"));
+        this.type = getLastSeg(itemsMap.get("machineType"));
         this.zone = getLastSeg(itemsMap.get("zone"));
         this.creationTime = convertStringToDate(itemsMap.get("creationTimestamp"));
         this.status = itemsMap.get("status");
