+++ src/main/java/ResourceDiscovery/AssetObjectsFactory.java
@@ -1,5 +1,7 @@
 package ResourceDiscovery;
 
+import ResourceDiscovery.AssetObjects.*;
+
 import java.util.Map;
 
 public class AssetObjectsFactory {
